day 11 part 2

+++ src/main.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 fn main() {
     let input = "475449 2599064 213 0 2 65 5755 51149";
     // let input = String::from("125 17");
-    let mut map: HashMap<String, i32> = HashMap::new();
+    let mut map: HashMap<String, i64> = HashMap::new();
     for i in input.split(" ").into_iter() {
         if map.contains_key(&String::from(i)) {
             let val = map.get(&String::from(i)).unwrap();
@@ -23,41 +23,41 @@ fn main() {
     }
 }
 
-fn process_input(in_map: &HashMap<String, i32>) -> HashMap<String, i32> {
+fn process_input(in_map: &HashMap<String, i64>) -> HashMap<String, i64> {
     let map = in_map.clone();
-    let mut out_map: HashMap<String, i32> = HashMap::new();
+    let mut out_map: HashMap<String, i64> = HashMap::new();
 
     for (key, val) in map.iter() {
         if key.eq("0") {
-            let count: i32 = *out_map.get(&String::from("1")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("1")).unwrap_or(&0);
             out_map.insert(String::from("1"), val.clone() + count);
         } else if key.eq("1") {
-            let count: i32 = *out_map.get(&String::from("2024")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("2024")).unwrap_or(&0);
             out_map.insert(String::from("2024"), val.clone() + count);
         } else if key.eq("20") {
-            let count: i32 = *out_map.get(&String::from("2")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("2")).unwrap_or(&0);
             out_map.insert(String::from("2"), val.clone() + count);
-            let count: i32 = *out_map.get(&String::from("0")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("0")).unwrap_or(&0);
             out_map.insert(String::from("0"), val.clone() + count);
         } else if key.eq("24") {
-            let count: i32 = *out_map.get(&String::from("2")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("2")).unwrap_or(&0);
             out_map.insert(String::from("2"), val.clone() + count);
-            let count: i32 = *out_map.get(&String::from("4")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("4")).unwrap_or(&0);
             out_map.insert(String::from("4"), val.clone() + count);
         } else if key.eq("2024") {
-            let count: i32 = *out_map.get(&String::from("20")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("20")).unwrap_or(&0);
             out_map.insert(String::from("20"), val.clone() + count);
-            let count: i32 = *out_map.get(&String::from("24")).unwrap_or(&0);
+            let count: i64 = *out_map.get(&String::from("24")).unwrap_or(&0);
             out_map.insert(String::from("24"), val.clone() + count);
         } else if key.len() % 2 == 0 {
             let (l, r) = split_stone(key.clone());
-            let count: i32 = *out_map.get(&l).unwrap_or(&0);
+            let count: i64 = *out_map.get(&l).unwrap_or(&0);
             out_map.insert(l, val.clone() + count);
-            let count: i32 = *out_map.get(&r).unwrap_or(&0);
+            let count: i64 = *out_map.get(&r).unwrap_or(&0);
             out_map.insert(r, val.clone() + count);
         } else if key.len() % 2 != 0 {
             let mul_op = mul_stone(key.clone());
-            let count: i32 = *out_map.get(&mul_op).unwrap_or(&0);
+            let count: i64 = *out_map.get(&mul_op).unwrap_or(&0);
             out_map.insert(mul_op, val.clone() + count);
         }
     }
